refactor: change css paths
