+++ src/pages/Games/SprintGame/CardSprint/CardEndGame/CardEndGame.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../common/style/abstract/_variables.scss';
+@import '../../../../../common/style/abstract/_variables.scss';
 .result {
     font: $h3-font;
 }
