+++ src/pages/Games/SprintGame/CardSprint/CardEndGame/CardEndGameInterface.ts
@@ -1,4 +1,4 @@
-import Word from '../../../../components/API/DictionaryAPI/Word';
+import Word from '../../../../../components/API/DictionaryAPI/Word';
 
 export default interface ICardInGame {
   mistakes: Word[];
