+++ src/pages/Games/SprintGame/CardSprint/CardEndGame/LinkToDictionary/LinkTodictionary.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../common/style/abstract/_variables.scss';
+@import '../../../../../../common/style/abstract/_variables.scss';
 .toDictionary {
     font: $paragraph-font;
 }
