+++ src/pages/Games/SprintGame/CardSprint/CardEndGame/WordList/WordList.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../common/style/abstract/_variables.scss';
+@import '../../../../../../common/style/abstract/_variables.scss';
 ul {
   padding: 0;
 }
