+++ src/pages/Games/SprintGame/CardSprint/CardEndGame/WordList/WordListHeader/WordListHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../common/style/abstract/_variables.scss';
+@import '../../../../../../../common/style/abstract/_variables.scss';
 .header {
     width: 110%;
     font: $h3-font;
