+++ src/pages/Games/SprintGame/CardSprint/CardWordGuessPart/CardWordGuessPart.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../common/style/abstract/_variables.scss';
+@import '../../../../../common/style/abstract/_variables.scss';
 .container {
     display: flex;
     justify-content: center;
