+++ src/pages/Games/SprintGame/CounterGame/CounterGame.module.scss
@@ -1,4 +1,4 @@
-@import '../../../common/style/abstract/_variables.scss';
+@import '../../../../common/style/abstract/_variables.scss';
 .container {
     position: absolute;
     width: 8rem;
