+++ src/pages/Games/SprintGame/ScoreGame/ScoreGame.module.scss
@@ -1,4 +1,4 @@
-@import '../../../common/style/abstract/_variables.scss';
+@import '../../../../common/style/abstract/_variables.scss';
 .container {
     font: $h3-font;
     text-align: center;
