+++ src/pages/Games/SprintGame/SprintDescription/SprintDescription.module.scss
@@ -1,4 +1,4 @@
-@import '../../../common/style/abstract/_variables.scss';
+@import '../../../../common/style/abstract/_variables.scss';
 .description {
     display: flex;
     flex-direction: column;
