+++ src/pages/Games/SprintGame/SprintDifficulty/Difficulty/Difficulty.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../common/style/abstract/_variables.scss';
+@import '../../../../../common/style/abstract/_variables.scss';
 .container {
   width: 8rem;
   height: 8rem;
