+++ src/pages/Games/SprintGame/SprintDifficulty/Difficulty/DifficultyInterfase.ts
@@ -1,4 +1,4 @@
-import { IDifficulty } from '../../../../common/constants/difficultyLevels';
+import { IDifficulty } from '../../../../../common/constants/difficultyLevels';
 
 export default interface IDifficultyGame {
   difficulty: string;
