+++ src/pages/Games/SprintGame/SprintDifficulty/SprintDifficulty.module.scss
@@ -1,4 +1,4 @@
-@import '../../../common/style/abstract/_variables.scss';
+@import '../../../../common/style/abstract/_variables.scss';
 .container {
   display: flex;
   justify-content: space-evenly;
