+++ src/pages/Games/SprintGame/SprintDifficulty/SprintDifficulty.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { DIFFICULTY } from '../../../common/constants/difficultyLevels';
+import { DIFFICULTY } from '../../../../common/constants/difficultyLevels';
 import Difficulty from './Difficulty/Difficulty';
 import classes from './SprintDifficulty.module.scss';
-import Button from '../../../components/Button/Button';
+import Button from '../../../../components/Button/Button';
 
 const SprintDifficulty = (props: { change: (difficultyLevel: number) => void }) => {
   const [active, setActive] = useState(false);
