+++ src/pages/Games/SprintGame/SprintGame/SprintGame.module.scss
@@ -1,6 +1,6 @@
 .background {
     position: relative;
-    background-image: url(../../../assets/images/Sprint/SprintBG.png);
+    background-image: url(../../../../assets/images/Sprint/SprintBG.png);
     background-repeat: none;
     background-size: cover;
 }
