+++ src/pages/Games/SprintGame/StartingPageSprint/StartingPageSprint.module.scss
@@ -1,5 +1,5 @@
 .background {
-  background-image: url(../../../assets/images/Sprint/SprintBG.png);
+  background-image: url(../../../../assets/images/Sprint/SprintBG.png);
   background-repeat: none;
   background-size: cover;
   width: 100vw;
